Convert ListView in allowed apps to RecyclerView
Closes #693
