+++ main/src/ui/res/layout/allowed_vpn_apps.xml
@@ -4,15 +4,17 @@
   -->
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:ignore="RtlCompat">
-
-    <ListView
-        android:id="@android:id/list"
+        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:orientation="vertical"
+        tools:ignore="RtlCompat">
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/app_recycler_view"
+        android:layout_width="match_parent"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         android:layout_height="match_parent"
         android:clipToPadding="false"
         android:drawSelectorOnTop="false"
